artemis-backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
Pillow
opencv-python
djoser
ensemble_boxes
ultralytics